[TIMOB-16386] Finished updating the hooks to use the updated detection system.

+++ mobileweb/cli/hooks/package_windows.js
@@ -5,21 +5,20 @@
  * See the LICENSE file for more information.
  */
 
-var path = require('path'),
-	fs = require('fs'),
-	spawn = require('child_process').spawn,
-	wrench = require('wrench'),
+var appc = require('node-appc'),
 	async = require('async'),
 	ejs = require('ejs'),
+	fs = require('fs'),
+	path = require('path'),
+	spawn = require('child_process').spawn,
 	uuid = require('node-uuid'),
-	appc = require('node-appc'),
-	i18n = appc.i18n(__dirname),
-	__ = i18n.__;
+	windows = require('titanium-sdk/lib/windows'),
+	wrench = require('wrench'),
+	__ = appc.i18n(__dirname).__;
 
 exports.cliVersion = '>=3.X';
 
 exports.init = function (logger, config, cli) {
-
 	if (process.platform != 'win32') {
 		return;
 	}
@@ -27,7 +26,6 @@ exports.init = function (logger, config, cli) {
 	cli.addHook('build.post.compile', {
 		priority: 8000,
 		post: function (build, finished) {
-
 			var target = cli.argv.target,
 				tiapp = build.tiapp,
 				displayName = target == 'wp8' ? __('Windows Phone 8') : __('Windows Store'),
@@ -39,61 +37,64 @@ exports.init = function (logger, config, cli) {
 			}
 
 			logger.info(__('Bundling Mobile Web app as a standalone %s app', displayName));
-			require('titanium-sdk/lib/' + target).detect(function (env) {
 
-				function run(command, callback) {
-					var proc = spawn('vcvarsall.bat', [ '&&' ].concat(command), {
-							cwd: path.join(env.visualStudioPath, 'VC')
-						});
-					proc.stdout.on('data', function (data) {
-						data.toString().split('\r\n').forEach(function (line) {
-							line = line.trim();
-							if (line.length) {
-								logger.trace(line);
-							}
-						});
-					});
-					proc.stderr.on('data', function (data) {
-						data.toString().split('\r\n').forEach(function (line) {
-							line = line.trim();
-							if (line.length) {
-								logger.error(line);
-							}
-						});
-					});
-					proc.on('error', function () {
-						logger.error('An unknown error occured while trying to spawn the external command');
-						callback(1);
-					});
-					proc.on('close', callback);
+			windows.detect(config, null, function (env) {
+				var vsInfo = windows.getSelectedVisualStudio(env);
+
+				if (!vsInfo) {
+					var validVersions = Object.keys(env.visualstudio).filter(function (v) { return env.visualstudio[v].supported; }).sort();
+					if (validVersions.length) {
+						logger.error(
+							__('Unable to find a suitable version of Microsoft Visual Studio.') + '\n' +
+							__('Manually select one of the following: %s', validVersions.join(', ')) + '\n' +
+							'  titanium config windows.visualstudio.selectedVersion <version>\n'
+						);
+					} else {
+						logger.error(
+							__('Unable to find a suitable version of Microsoft Visual Studio.') + '\n' +
+							__('You can install it from %s.', '__http://appcelerator.com/visualstudio__') + '\n'
+						);
+					}
+					return finished(1);
 				}
 
-				if (env.issues.length) {
-					logger.error(__('There were issues detected with the %s development environment setup. ' +
-						'Please run "titanium info" for more information', displayName));
-					process.exit(1);
+				function makePipe(log) {
+					return function (data) {
+						data.toString().split(/\r\n|\n/).forEach(function (line) {
+							line = line.trim();
+							line.length && log(line);
+						});
+					};
+				}
+
+				function run(command, callback) {
+					var child = spawn(vsInfo.vcvarsall, ['&&'].concat(command), {
+						cwd: path.join(vsInfo.path, 'VC')
+					});
+					child.stdout.on('data', makePipe(logger.trace));
+					child.stderr.on('data', makePipe(logger.error));
+					child.on('close', callback);
 				}
 
 				// Check if the certificate has been generated yet or not
 				if (target == 'winstore' && !fs.existsSync(certificatePathRoot + '.pfx')) {
-
-					logger.info(__('No code signing certificate for Windows Store applications was found, ' +
-						'creating and installing one now. ' +
-						'The certificate will be installed in the local machine\'s Trusted Root Certificate Authorities certificate store. ' +
-						'When prompted for a password, leave everything blank.'));
+					logger.info(__('No code signing certificate for Windows Store applications was found, creating and installing one now.'));
+					logger.info(__("The certificate will be installed in the local machine's Trusted Root Certificate Authorities certificate store."));
+					logger.info(__('When prompted for a password, leave everything blank.'));
 
 					async.series([
-
 						// Create the certificate and install it in the root trusted certificates
 						function (next) {
 							var cmd = [
-									'powershell.exe',
-									'-command',
-									path.join(__dirname, '..', '..', '..', 'node_modules', 'titanium-sdk', 'bin', 'winstore_create_cert.ps1') + ' ' + tiapp.id + ' "' + certificatePathRoot + '"'
-								];
+								config.get('windows.executables.powershell', 'powershell'),
+								'-command',
+								path.resolve(__dirname, '..', '..', '..', 'node_modules', 'titanium-sdk', 'bin', 'winstore_create_cert.ps1'),
+								tiapp.id,
+								certificatePathRoot
+							];
 
 							logger.info(__('Generating the code signing certificate'));
-							logger.debug(__('Running: %s', ('"' + cmd.join('" "') + '"').cyan));
+							logger.debug(__('Running: %s', cmd.join(' ').cyan));
 
 							run(cmd, function (code) {
 								if (code) {
@@ -107,25 +108,32 @@ exports.init = function (logger, config, cli) {
 
 						// Package the certificate as pfx
 						function (next) {
-							logger.debug(__('Exporting the certificate to PKCS#12'));
-							run([ 'Pvk2Pfx',
-									'/pvk', certificatePathRoot + '.pvk',
-									'/spc', certificatePathRoot + '.cer',
-									'/pfx', certificatePathRoot + '.pfx'
-								], function (code) {
-									logger.debug(__('Removing intermediate certificate files'));
-									fs.unlinkSync(certificatePathRoot + '.pvk');
-									fs.unlinkSync(certificatePathRoot + '.cer');
-									if (code) {
-										logger.error(__('There were errors exporting the code signing certificate'));
-									} else {
-										logger.debug(__('Finished creating the code signing certificate'));
-									}
-									next(code);
-								});
+							var cmd = [
+								config.get('windows.executables.pvk2pfx', 'Pvk2Pfx'),
+								'/pvk', certificatePathRoot + '.pvk',
+								'/spc', certificatePathRoot + '.cer',
+								'/pfx', certificatePathRoot + '.pfx'
+							];
+
+							logger.info(__('Exporting the certificate to PKCS#12'));
+							logger.debug(__('Running: %s', cmd.join(' ').cyan));
+
+							run(cmd, function (code) {
+								logger.debug(__('Removing intermediate certificate files'));
+								fs.existsSync(certificatePathRoot + '.pvk') && fs.unlinkSync(certificatePathRoot + '.pvk');
+								fs.existsSync(certificatePathRoot + '.cer') && fs.unlinkSync(certificatePathRoot + '.cer');
+								if (code) {
+									logger.error(__('There were errors exporting the code signing certificate'));
+								} else {
+									logger.debug(__('Finished creating the code signing certificate'));
+								}
+								next(code);
+							});
 						}
 					], function (err) {
-						if (!err) {
+						if (err) {
+							finished(err);
+						} else {
 							packageApp();
 						}
 					});
@@ -134,64 +142,66 @@ exports.init = function (logger, config, cli) {
 				}
 
 				function packageApp() {
-
 					var source = path.resolve(build.buildDir),
-						destination = path.resolve(path.join(source, '..', 'mobileweb-' + target)),
+						destination = path.resolve(source, '..', 'mobileweb-' + target),
 						version = tiapp.version,
-						templateData = target == 'wp8' ? {
-								projectName: tiapp.id,
-								projectDisplayName: tiapp.name,
-								projectGUID: tiapp.guid || uuid.v4(),
-								assemblyGUID: uuid.v4(),
-								publisherGUID: cli.argv['wp8-publisher-guid'],
-								company: 'not specified', // Hopefully we can support this some day
-								projectDescription: tiapp.description || '',
-								author: tiapp.publisher,
-								copyright: tiapp.copyright || 'Copyright © ' + new Date().getFullYear(),
-								appFiles: []
-							} : {
-								projectName: tiapp.id,
-								projectDisplayName: tiapp.name,
-								projectGUID: tiapp.guid || uuid.v4(),
-								projectDescription: tiapp.description || '',
-								author: config.user.name || 'Titanium',
-								appFiles: [],
-								visualStudioVersion: env.visualStudioVersion,
-								certificatePath: certificatePathRoot + '.pfx'
-							},
+						templateData = {
+							// general
+							projectName: tiapp.id,
+							projectDisplayName: tiapp.name,
+							projectGUID: tiapp.guid || uuid.v4(),
+							projectDescription: tiapp.description || '',
+							author: tiapp.publisher || config.get('user.name') || 'Titanium',
+							appFiles: [],
+
+							// windows phone specific
+							assemblyGUID: uuid.v4(),
+							publisherGUID: cli.argv['wp8-publisher-guid'],
+							company: 'not specified', // Hopefully we can support this some day
+							copyright: tiapp.copyright || ('Copyright © ' + new Date().getFullYear()),
+
+							// windows store specific
+							visualStudioVersion: env.visualStudioVersion,
+							certificatePath: certificatePathRoot + '.pfx'
+						},
 						templateDir = path.join(__dirname, '..', '..', 'templates', 'packages', target),
 						filenameReplacementRegex = /\{\{ProjectName\}\}/g,
-						templateFiles = target == 'wp8' ? [
-								'{{ProjectName}}.sln',
-								path.join('{{ProjectName}}', '{{ProjectName}}.csproj'),
-								path.join('{{ProjectName}}', 'MainPage.xaml'),
-								path.join('{{ProjectName}}', 'MainPage.xaml.cs'),
-								path.join('{{ProjectName}}', 'LocalizedStrings.cs'),
-								path.join('{{ProjectName}}', 'App.xaml'),
-								path.join('{{ProjectName}}', 'App.xaml.cs'),
-								path.join('{{ProjectName}}', 'Resources', 'AppResources.Designer.cs'),
-								path.join('{{ProjectName}}', 'Properties', 'AssemblyInfo.cs'),
-								path.join('{{ProjectName}}', 'Properties', 'WMAppManifest.xml')
-							] : [
-								'{{ProjectName}}.sln',
-								path.join('{{ProjectName}}', '{{ProjectName}}.jsproj'),
-								path.join('{{ProjectName}}', 'package.appxmanifest')
-							],
+						templateFiles,
 						appFiles = templateData.appFiles,
-						buildProcess,
 						versionFormatRegex = /^[0-9]+\.[0-9]+\.[0-9]+\.[0-9]+$/;
 
-					// Validate and format the version to Major.Minor.Build.Revision format necessary for MS build systems
+					if (target == 'wp8') {
+						templateFiles = [
+							'{{ProjectName}}.sln',
+							path.join('{{ProjectName}}', '{{ProjectName}}.csproj'),
+							path.join('{{ProjectName}}', 'MainPage.xaml'),
+							path.join('{{ProjectName}}', 'MainPage.xaml.cs'),
+							path.join('{{ProjectName}}', 'LocalizedStrings.cs'),
+							path.join('{{ProjectName}}', 'App.xaml'),
+							path.join('{{ProjectName}}', 'App.xaml.cs'),
+							path.join('{{ProjectName}}', 'Resources', 'AppResources.Designer.cs'),
+							path.join('{{ProjectName}}', 'Properties', 'AssemblyInfo.cs'),
+							path.join('{{ProjectName}}', 'Properties', 'WMAppManifest.xml')
+						];
+					} else {
+						templateFiles = [
+							'{{ProjectName}}.sln',
+							path.join('{{ProjectName}}', '{{ProjectName}}.jsproj'),
+							path.join('{{ProjectName}}', 'package.appxmanifest')
+						];
+					}
+
+					// validate and format the version to Major.Minor.Build.Revision format necessary for MS build systems
 					if (!version) {
 						version = '0.1.0.0';
-					} else if (!versionFormatRegex.exec(version)) {
+					} else if (!versionFormatRegex.test(version)) {
 						version = version.match(/^[0-9]+(\.[0-9]+)*/);
 						if (!version) {
 							logger.warn(__('Invalid project version number %s, setting to 0.1.0.0', tiapp.version));
 							version = '0.1.0.0';
 						} else {
 							version = version[0];
-							while(!versionFormatRegex.exec(version)) {
+							while (!versionFormatRegex.test(version)) {
 								version = version + '.0';
 							}
 							logger.info(__('Project version number will be converted to %s for compatibility with Visual Studio', version));
@@ -251,37 +261,19 @@ exports.init = function (logger, config, cli) {
 					});
 
 					// Compile the app
+					var cmd = [
+						'MSBuild',
+						'/m',
+						'/p:configuration=' + (cli.argv['deploy-type'] == 'production' ? 'Release' : 'Debug'),
+						path.join(destination, tiapp.id + '.sln')
+					];
 					logger.info(__('Building the %s Visual Studio project', displayName));
-					buildProcess = spawn('vcvarsall.bat', [
-							'&&',
-							'MSBuild',
-							'/m',
-							'/p:configuration=' + (cli.argv['deploy-type'] == 'production' ? 'Release' : 'Debug'),
-							path.join(destination, tiapp.id + '.sln')
-						], {
-							cwd: path.join(env.visualStudioPath, 'VC')
-						});
-					buildProcess.stdout.on('data', function (data) {
-						data.toString().split('\r\n').forEach(function (line) {
-							if (line.length) {
-								logger.trace(line);
-							}
-						});
-					});
-					buildProcess.stderr.on('data', function (data) {
-						data.toString().split('\r\n').forEach(function (line) {
-							if (line.length) {
-								logger.error(line);
-							}
-						});
-					});
-					buildProcess.on('close', function (code) {
-						if (code) {
-							finished(code);
-						} else {
+					logger.debug(__('Running: %s', cmd.join(' ').cyan));
+					run(cmd, function (code) {
+						if (!code) {
 							logger.info(__('Finished building the application'));
-							finished();
 						}
+						finished(code);
 					});
 				}
 			});
